dihedral: Fix performance issue for Frechet mean
The metric function was compiled by numba on each call to the outer function.
Seems that this is the default behaviour in numba.
https://github.com/numba/numba/issues/1810

Moved to outer so that it's compiled once. Speedup is ~100x.

Added a benchmark test to see the performnace of this function.
After the change it's indeed improved by numba.