pgroup bugfix: structures with multiple same-entity chains
In cases where there were multiple chains for the same structure which
are the same entity, and the pgroup was initialized with a chain other
than the first, there would be a bug which would cause us to not find
REFERENCE matches.

For example, 6R67:B.
Since we get entitiy from the PDB query, we got 6R67 in our query list.
This was translated into 6R67:A when the reference prec was created.
Therefore, there were no "REFERENCE"-type matches, since 6R67:B!=6R67:A.